+++ src/main/java/de/uni_kl/cs/discodnc/feedforward/AnalysisConfig.java
@@ -26,7 +26,7 @@
  *
  */
 
-package de.uni_kl.cs.discodnc.nc;
+package de.uni_kl.cs.discodnc.feedforward;
 
 import java.util.Collections;
 import java.util.HashSet;
