--- src/main/java/de/uni_kl/cs/discodnc/nc/analyses/PmooResults.java
@@ -1,65 +1,0 @@
-/*
- * This file is part of the Disco Deterministic Network Calculator.
- *
- * Copyright (C) 2005 - 2007 Frank A. Zdarsky
- * Copyright (C) 2008 - 2010 Andreas Kiefer
- * Copyright (C) 2011 - 2018 Steffen Bondorf
- * Copyright (C) 2017+ The DiscoDNC contributors
- *
- * Distributed Computer Systems (DISCO) Lab
- * University of Kaiserslautern, Germany
- *
- * http://discodnc.cs.uni-kl.de
- *
- *
- * The Disco Deterministic Network Calculator (DiscoDNC) is free software;
- * you can redistribute it and/or modify it under the terms of the 
- * GNU Lesser General Public License as published by the Free Software Foundation; 
- * either version 2.1 of the License, or (at your option) any later version.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
- *
- */
-
-package de.uni_kl.cs.discodnc.nc.analyses;
-
-import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
-import de.uni_kl.cs.discodnc.curves.ServiceCurve;
-import de.uni_kl.cs.discodnc.nc.AnalysisResults;
-import de.uni_kl.cs.discodnc.network.Server;
-import de.uni_kl.cs.discodnc.numbers.Num;
-
-import java.util.Map;
-import java.util.Set;
-
-public class PmooResults extends AnalysisResults {
-    protected Set<ServiceCurve> betas_e2e;
-
-    protected PmooResults() {
-    }
-
-    protected PmooResults(Num delay_bound, Num backlog_bound, Set<ServiceCurve> betas_e2e,
-                          Map<Server, Set<ArrivalCurve>> map__server__alphas) {
-
-        super(delay_bound, backlog_bound, map__server__alphas);
-
-        this.betas_e2e = betas_e2e;
-    }
-
-    @Override
-    protected void setDelayBound(Num delay_bound) {
-        super.setDelayBound(delay_bound);
-    }
-
-    @Override
-    protected void setBacklogBound(Num backlog_bound) {
-        super.setBacklogBound(backlog_bound);
-    }
-}
