+++ src/main/java/de/uni_kl/cs/discodnc/network/Server.java
@@ -31,7 +31,7 @@ package de.uni_kl.cs.discodnc.network;
 import de.uni_kl.cs.discodnc.curves.Curve;
 import de.uni_kl.cs.discodnc.curves.MaxServiceCurve;
 import de.uni_kl.cs.discodnc.curves.ServiceCurve;
-import de.uni_kl.cs.discodnc.nc.AnalysisConfig.Multiplexing;
+import de.uni_kl.cs.discodnc.feedforward.AnalysisConfig.Multiplexing;
 
 public class Server {
     private int id;
